# [MRG] MNT Initialize histograms in parallel and don't call np.zero in Hist-GBDT