--- a/doc/whats_new/v0.24.rst
+++ b/doc/whats_new/v0.24.rst
@@ -198,8 +198,9 @@ Changelog
 - |Efficiency| Histogram initialization is now done in parallel in
   :class:`ensemble.HistGradientBoostingRegressor` and
   :class:`ensemble.HistGradientBoostingClassifier` which results in speed
-  improvement for problems that build a lot of nodes.
-  :pr:`18341` by `Olivier Grisel`_, `Nicolas Hug`_, `Thomas Fan`_.
+  improvement for problems that build a lot of nodes on multicore machines.
+  :pr:`18341` by `Olivier Grisel`_, `Nicolas Hug`_, `Thomas Fan`_, and
+  :user:`Egor Smirnov <SmirnovEgorRu>`.
 
 - |API|: The parameter ``n_classes_`` is now deprecated in
   :class:`ensemble.GradientBoostingRegressor` and returns `1`.