To keep things more unitary, this PR extracts out the histogram initialization from https://github.com/scikit-learn/scikit-learn/pull/18242 which was already proposed in https://github.com/scikit-learn/scikit-learn/pull/14392

Memory usage is the same as in master but runs in 40s instead of 45s (across many runs) on the following benchmarks on my laptop (4 threads):


```py
from sklearn.datasets import make_classification
from sklearn.experimental import enable_hist_gradient_boosting
from sklearn.ensemble import HistGradientBoostingClassifier
from memory_profiler import memory_usage

X, y = make_classification(n_classes=5,
                           n_samples=3_000,
                           n_features=400,
                           random_state=0,
                           n_clusters_per_class=1,
                           n_informative=5)

hgb = HistGradientBoostingClassifier(
    max_iter=100,
    max_leaf_nodes=256,
    learning_rate=.1,
    random_state=0,
    verbose=1,
)

mems = memory_usage((hgb.fit, (X, y)))
print(f"{max(mems):.2f}, {max(mems) - min(mems):.2f} MB")
```

Other benchmarks welcome, especially one with 88 hyperthreads as in https://github.com/scikit-learn/scikit-learn/pull/14392#issuecomment-512345250.

CC @ogrisel @thomasjpfan