@ogrisel 2020-09-04
Ok let me re-run a quick benchmark with many threads on this PR.