@ogrisel 2020-09-04
On a machine with 36 physical cores (72 HyperThreads):

- commandline:

```
$ OMP_NUM_THREADS=36 nice  python benchmarks/bench_hist_gradient_boosting_higgsboson.py --n-trees 100 --n-leaf-nodes 255
```

- master with 

```
Fit 100 trees in 39.437 s, (25500 total leaves)                                                                                                                               
Time spent computing histograms: 17.777s                                                                                                                                      
Time spent finding best splits:  1.128s                                                                                                                                       
Time spent applying splits:      4.035s                                                                                                                                       
Time spent predicting:           1.743s                                                                                                                                       
fitted in 39.451s                                                                                                                                                             
predicted in 2.488s, ROC AUC: 0.8280, ACC: 0.7466
```

- this branch (`hist_parallel`):

```
Fit 100 trees in 39.656 s, (25500 total leaves)
Fit 100 trees in 38.755 s, (25500 total leaves)
Time spent computing histograms: 16.972s
Time spent finding best splits:  1.147s
Time spent applying splits:      4.027s
Time spent predicting:           1.753s
fitted in 38.773s
predicted in 2.950s, ROC AUC: 0.8278, ACC: 0.7462
```

so no significant improvement (I re-rerun the benchmark several times and observe approximately +/- 0.5s in total fit time for both branches

I was almost alone on the machine when running the benchmark.

For information lightgbm yields with 36 threads on the same machine:

```
fitted in 17.750s
predicted in 1.265s, ROC AUC: 0.8268, ACC: 0.7453
```