@ogrisel 2020-09-04
Without limiting the number of threads to the number of physical cores (no `OMP_NUM_THREADS`):

- master

```
Fit 100 trees in 51.788 s, (25500 total leaves)                                                                                                                               
Time spent computing histograms: 23.774s                                                                                                                                      
Time spent finding best splits:  2.760s                                                                                                                                       
Time spent applying splits:      6.218s                                                                                                                                       
Time spent predicting:           1.838s                               
fitted in 51.808s                                                     
predicted in 2.669s, ROC AUC: 0.8283, ACC: 0.7468
```

- this branch:

```
Fit 100 trees in 46.983 s, (25500 total leaves)
Time spent computing histograms: 21.686s
Time spent finding best splits:  2.015s
Time spent applying splits:      5.202s
Time spent predicting:           1.947s
fitted in 47.001s
predicted in 2.621s, ROC AUC: 0.8279, ACC: 0.7463
```

- lightgbm

```
fitted in 20.194s
predicted in 1.118s, ROC AUC: 0.8274, ACC: 0.7459
```

so maybe when the number of threads is larger than the number of physical CPU cores (oversubscription), this branch make it possible to mitigate some of the perf degradation cause by over-subscription. But this is probably in the noise because I get +/- 2s variations in that regime.

So in conclusion, no significant performance impact on many threads machines.