@NicolasHug 2020-09-04
Thanks for the benchmarks @ogrisel. 

Just to make sure, did you use the snippet provided here? It's different from the other ones, as I designed it  to build lots of histograms.

I tried on my desktop computer with 16 threads and the execution time drops from 28s (master) to 10s (this PR). So that's quite a  significant improvement in this case.

Since you did not observe a regression in your benchmarks, I'd say we still have a strong incentive to merge this as a stand-alone.