@NicolasHug 2020-09-04
My CPU with 16 threads has only 8 cores. Using OMP_NUM_THREADS=8, across multiple runs:

- 22s on master
- 10s on this PR

So in my case I observe a significant improvement *not just* in the case of over-subscription.