@ogrisel 2020-09-04
For the code snippet this yields a huge improvement:

- 72 threads (on a 36 cores machine) / master: 47s
- 72 threads (on a 36 cores machine) / this branch: 18s

- 36 threads (on a 36 cores machine) / master: 35s
- 36 threads (on a 36 cores machine) / this branch: 14s