@lorentzenchr 2020-09-04
On macOS without openMP, no regression as well.
- master
  ```
  Fit 500 trees in 42.757 s, (47303 total leaves)
  Time spent computing histograms: 17.263s
  Time spent finding best splits:  19.739s
  Time spent applying splits:      1.075s
  Time spent predicting:           0.029s
  164.78, 34.19 MB
  ```
- this PR
  ```
  Fit 500 trees in 43.890 s, (47303 total leaves)
  Time spent computing histograms: 16.605s
  Time spent finding best splits:  21.601s
  Time spent applying splits:      1.024s
  Time spent predicting:           0.029s
  164.91, 35.30 MB
  ```