@ogrisel 2020-09-04
I have updated https://github.com/scikit-learn/scikit-learn/pull/18341#issuecomment-687138782 to also add bench results for #18242's branch with parallel zero init + histogrampool.

The histogrampool is performance neutral compared to this branch.