@NicolasHug 2020-09-04
@ogrisel I believe this PR may close https://github.com/scikit-learn/scikit-learn/issues/14306, WDYT?