@ogrisel 2020-09-04
Thanks for crediting Egor indeed.