@ogrisel 2020-09-04
> @ogrisel I believe this PR may close #14306, WDYT?

I am ambivalent. But the memory usage problem @thomasjpfan might be too rare to justify the added code complexity. Let's wait for his opinion. In the mean time we can already merge this one.