@NicolasHug 2020-09-04
I believe  #14306 isn't about memory but about thread scalability. In particular @SmirnovEgorRu  had identified that `np.zeros` might be a problem