@ogrisel 2020-09-04
Oops I merged without waiting for the CI to be green but I am confident this will not crash the build on master. Fingers crossed.