@ogrisel 2020-09-04
> I believe #14306 isn't about memory but about thread scalability. In particular @SmirnovEgorRu had identified that np.zeros might be a problem

Indeed, I was mistaken. It's an improvement but it still not as good as lightgbm in terms of scalability. But I agree we should update the benchmark results in that issue to reflect the new master.